@@ -1,64 +1,37 @@
 <template>
-  <div :class="{'has-logo':showLogo}">
-    <logo v-if="showLogo" :collapse="isCollapse" :organizations="organizations" />
-    <el-scrollbar wrap-class="scrollbar-wrapper">
-      <el-menu
-        :default-active="activeMenu"
-        :collapse="isCollapse"
-        :background-color="variables.menuBg"
-        :text-color="variables.menuText"
-        :unique-opened="false"
-        :active-text-color="variables.menuActiveText"
-        :collapse-transition="false"
-
-        mode="vertical"
-      >
-        <sidebar-item v-for="route in permission_routes" :key="route.path" :item="route" :base-path="route.path" />
-      </el-menu>
-    </el-scrollbar>
+  <div :class="{ 'has-logo': sidebarLogo }">
+    <!--混合布局-->
+    <div class="flex w-full" v-if="layout == 'mix'">
+      <SidebarLogo v-if="sidebarLogo" :collapse="!appStore.sidebar.opened" />
+      <SidebarMixTopMenu class="flex-1" />
+      <NavbarRight />
+    </div>
+    <!--左侧布局 || 顶部布局 -->
+    <template v-else>
+      <SidebarLogo v-if="sidebarLogo" :collapse="!appStore.sidebar.opened" />
+      <el-scrollbar>
+        <SidebarMenu :menu-list="permissionStore.routes" base-path="" />
+      </el-scrollbar>
+      <NavbarRight v-if="layout === 'top'" />
+    </template>
   </div>
 </template>
 
-<script>
-import { mapGetters } from 'vuex'
-import Logo from './Logo'
-import SidebarItem from './SidebarItem'
-import variables from '@/styles/variables.scss'
+<script setup lang="ts">
+import { useSettingsStore, usePermissionStore, useAppStore } from "@/store";
 
-export default {
-  components: { SidebarItem, Logo },
-  computed: {
-    ...mapGetters([
-      'permission_routes',
-      'sidebar'
-    ]),
-    activeMenu() {
-      const route = this.$route
-      const { meta, path } = route
-      // if set path, the sidebar will highlight the path you set
-      if (meta.activeMenu) {
-        return meta.activeMenu
-      }
-      return path
-    },
-    showLogo() {
-      return this.$store.state.settings.sidebarLogo
-    },
-    variables() {
-      return variables
-    },
-    isCollapse() {
-      return !this.sidebar.opened
-    },
-    organizations() {
-      return [
-        {
-          id: 1,
-          name: '测试',
-          url: 'http://127.0.0.1:9527'
-        }
-      ]
-    }
+const appStore = useAppStore();
+const settingsStore = useSettingsStore();
+const permissionStore = usePermissionStore();
+
+const sidebarLogo = computed(() => settingsStore.sidebarLogo);
+const layout = computed(() => settingsStore.layout);
+</script>
+
+<style lang="scss" scoped>
+.has-logo {
+  .el-scrollbar {
+    height: calc(100vh - $navbar-height);
   }
 }
-</script>
+</style>
